Separated Out Classes From "jumper.py" File. - WordGuess is now in it's own file - PlayerInput is not in it's own file
